cmd/agent/requirements.txt
boto3==1.35.80
langgraph==0.2.62
langchain-aws==0.2.10
langchain-core==0.3.28
langchain-community==0.3.13
pydantic==2.10.6
pyyaml==6.0.2
